Restructure backend modules
